backend/requirements.txt
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pytz==2023.3
django-phonenumber-field==7.2.0
requests== 2.31.0
celery==5.3.6
redis==5.0.1